Fix autoload issues causing multiple instance of the AutoLoad system to instanciate at once. Improved separation of concerns
